Explicit unbox
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1719963 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/org/apache/jmeter/report/processor/AggregateConsumer.java
@@ -99,7 +99,7 @@ public class AggregateConsumer extends AbstractSampleConsumer {
      */
     @Override
     public void consume(Sample sample, int channel) {
-        aggregator.addValue(selector.select(sample));
+        aggregator.addValue(selector.select(sample).doubleValue());
         super.produce(sample, channel);
     }
